fix bugs

+++ highscore_test.py
@@ -4,18 +4,18 @@ import shutil
 import pytest
 from game.highscore import Highscore
 
-os.chdir('game.test_data')
 
 def clear_dir():
     for filename in os.listdir(os.getcwd()):
         file_path = os.path.join(os.getcwd(), filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print('Failed to delete %s. Reason: %s' % (file_path, e))
+        if 'data' in filename:
+            try:
+                if os.path.isfile(file_path) or os.path.islink(file_path):
+                    os.unlink(file_path)
+                elif os.path.isdir(file_path):
+                    shutil.rmtree(file_path)
+            except Exception as e:
+                print('Failed to delete %s. Reason: %s' % (file_path, e))
 
 def test_adding():
     h = Highscore()
